FAI-1232-Import-Answers-From-Previous-Application

+++ src/SFA.DAS.CandidateAccount.Application/Application/Commands/DeleteQualification/DeleteQualificationCommandHandler.cs
@@ -1,12 +1,27 @@
 using MediatR;
+using SFA.DAS.CandidateAccount.Data.Application;
 using SFA.DAS.CandidateAccount.Data.Qualification;
+using SFA.DAS.CandidateAccount.Domain.Application;
 
 namespace SFA.DAS.CandidateAccount.Application.Application.Commands.DeleteQualification;
 
-public class DeleteQualificationCommandHandler(IQualificationRepository qualificationRepository) : IRequestHandler<DeleteQualificationCommand, Unit>
+public class DeleteQualificationCommandHandler(IQualificationRepository qualificationRepository, IApplicationRepository applicationRepository) : IRequestHandler<DeleteQualificationCommand, Unit>
 {
     public async Task<Unit> Handle(DeleteQualificationCommand request, CancellationToken cancellationToken)
     {
+        var application = await applicationRepository.GetById(request.ApplicationId);
+
+        if (application == null || application.CandidateId != request.CandidateId)
+        {
+            throw new InvalidOperationException($"Application {request.ApplicationId} not found");
+        }
+
+        if (application.QualificationsStatus is (short)SectionStatus.PreviousAnswer)
+        {
+            application.QualificationsStatus = (short)SectionStatus.InProgress;
+            await applicationRepository.Update(application);
+        }
+
         await qualificationRepository.DeleteCandidateApplicationQualificationById(request.CandidateId,
             request.ApplicationId, request.Id);
